修复小数不能存的bug

+++ src/main/java/com/lingfei/admin/service/MoneyManageService.java
@@ -45,7 +45,7 @@ public interface MoneyManageService {
      * @param id
      * @return
      */
-    int updateBalance(float balance,int id);
+    int updateBalance(double balance,int id);
 
     /**
      * 根据id更新数据
